Use `renameat` etc. to improve atomicity
In `replace_atomic`, open the parent directories, use `renameat` to do
the rename, and then do `fsync` on the open parent directory file
descriptors, to be robust in case the parent directories are
concurrently renamed.

In `move_atomic`, use `renameat2` with `RENAME_NOREPLACE` when
available, and `linkat`+`unlinkat` otherwise, and do the `fsync`s on
the parent directories as in `replace_atomic`.

This adds a dependency on rustix, which requires an msrv of 1.48.

Fixes #28.